+++ src/core/Logger.hpp
@@ -12,5 +12,5 @@ public:
     static void error(std::string_view message);
 
 private:
-    static void writeLevel(std::string_view level, std::string_view message);
+    static void _writeLevel(std::string_view level, std::string_view message);
 };