+++ src/systems/RenderSystem.hpp
@@ -2,10 +2,12 @@
 
 #include "systems/System.hpp"
 
+#include "rendering/Scene.hpp"
+
 class RenderSystem : public System {
 public:
     using System::System;
     ~RenderSystem() override = default;
 
-    virtual void render() = 0;
+    virtual void render(const Scene& scene) = 0;
 };